TWITTER/v1/requirements.txt
tweepy
pillow
textwrap
schedule
request
regex